Waargonaut.Decode.Internal

Description

Internal types and functions for building Decoder infrastructure.

Documentation

newtype CursorHistory' i

Track the history of the cursor as we move around the zipper.

It is indexed over the type of the index used to navigate the zipper.

Constructors

CursorHistory'

Fields

unCursorHistory' :: Seq (ZipperMove, i)

ppCursorHistory :: Show i => CursorHistory' i -> Doc a

newtype DecodeResultT i e f a

The general structure used to maintain the history of the moves around the zipper, as well as handle the decoding or movement errors that may occur. This structure is generalised of the inner f to allow you to interleave the decoding with your own actions. As well as the error type e so that you may provide your own error type.

If you use the provided Decode module then you probably won't need to care about this type. It is provided so that you're not limited to how we decide you should be running your decoder.

Constructors

DecodeResultT

Fields

runDecodeResult :: ExceptT e (StateT (CursorHistory' i) f) a

newtype Decoder' c i e f a

Wrapper type to describe a Decoder from something that has a Jsonish value c, to some representation of a.

Constructors

Decoder'

Fields

runDecoder' :: c -> DecodeResultT i e f a

withCursor' :: (c -> DecodeResultT i e f a) -> Decoder' c i e f a

Helper function for constructing a Decoder'.

This function is used by the implemented decoders to simplify constructing a more specific Decoder type.

withCursor' $ curs ->
  ...
  ...

runDecoderResultT :: (Num i, Monad f) => DecodeResultT i DecodeError f a -> f (Either (DecodeError, CursorHistory' i) a)

Execute a given DecoderResultT.

If you're building your own decoder structure, this function will take care of the CursorHistory' and error handling (via ExceptT).

try :: MonadError e m => m a -> m (Maybe a)

Attempt a Decoder action that might fail and return a Maybe value instead.

recordZipperMove :: MonadState (CursorHistory' i) m => ZipperMove -> i -> m ()

Record a move on the zipper and the index of the position where the move occured.

Generalised Decoder Functions

null' :: AsJType a ws a => a -> Maybe ()

Try to decode a null value from some Json value

int' :: AsJType a ws a => a -> Maybe Int

Try to decode an Int from some Json value

text' :: AsJType a ws a => a -> Maybe Text

Try to decode a Text value from some Json or value.

string' :: AsJType a ws a => a -> Maybe String

Try to decode a String value from some Json or value.

unboundedChar' :: AsJType a ws a => a -> Maybe Char

Decoder for a Haskell Char value whose values represent Unicode (or equivalently ISO/IEC 10646) characters

boundedChar' :: AsJType a ws a => a -> Maybe Char

Decoder for a Char value that cannot contain values in the range U+D800 to U+DFFF. This decoder will fail if the Char is outside of this range.

bool' :: AsJType a ws a => a -> Maybe Bool

Try to decode a Bool from some Json value

array' :: AsJType a ws a => (a -> Maybe b) -> a -> [b]

Combined with another decoder function f, try to decode a list of a values.

array' int' :: Json -> [Int]

integral' :: (Bounded i, Integral i, AsJType a ws a) => a -> Maybe i

Try to decode a bounded 'Integral n => n' value from some Json value.

scientific' :: AsJType a ws a => a -> Maybe Scientific

Try to decode a Scientific value from some Json or value.

objTuples' :: (Applicative f, AsJType a ws a) => (JString -> f k) -> (a -> f b) -> a -> f [(k, b)]

Try to decode a JSON Object into it's representative list of tuples '(key, value)'. The JSON RFC does not specify that an object must contain unique keys. We do not enforce unique keys during the decoding process and leave it to the user to decide if, and how, they would like to handle this situation.

foldCursor' :: Monad f => b -> (b -> a -> b) -> (c -> DecodeResultT i e f c) -> Decoder' c i e f a -> c -> DecodeResultT i e f b

Generalised moving decoder function.

Starting from the given cursor position, try to move in the direction specified by the given cursor function. Attempting to decode each item at each position using the given Decoder, until the movement is unsuccessful.

The following could be used to leverage the Snoc instance of '[]' to build '[Int]'.

intList :: Monad f => JCurs -> DecodeResult f [Int]
intList = directedConsumption' snoc moveRight1 int

JSON Object to Map Functions

mapKeepingF :: (Ord k, Applicative f, AsJType a ws a) => (t -> Maybe v -> Maybe v) -> (JString -> f k) -> (a -> f t) -> a -> f (Map k v)

Provide a generalised and low level way of turning a JSON object into a Map, without enforcing a choice of how we select keys.

mapKeepingFirst :: (Ord k, Applicative f, AsJType a ws a) => (JString -> f k) -> (a -> f b) -> a -> f (Map k b)

Turn a JSON object into a Map by keeping the *first* occurence of any duplicate keys that are encountered.

mapKeepingLast :: (Ord k, Applicative f, AsJType a ws a) => (JString -> f k) -> (a -> f b) -> a -> f (Map k b)

Turn a JSON object into a Map by keeping the *last* occurence of any duplicate keys that are encountered.

Re-exports

class AsDecodeError r where

Minimal complete definition

_DecodeError

Methods

_DecodeError :: Prism' r DecodeError

_ConversionFailure :: Prism' r Text

_KeyDecodeFailed :: Prism' r Text

_KeyNotFound :: Prism' r Text

_FailedToMove :: Prism' r ZipperMove

_NumberOutOfBounds :: Prism' r JNumber

_InputOutOfBounds :: Prism' r Word64

_ParseFailed :: Prism' r Text

data DecodeError

Set of errors that may occur during the decode phase.

Constructors

ConversionFailure Text

KeyDecodeFailed Text

KeyNotFound Text

FailedToMove ZipperMove

NumberOutOfBounds JNumber

InputOutOfBounds Word64

ParseFailed Text

data ZipperMove

Set of moves that may be executed on a zipper.

Constructors

U

D

DAt Text

Item Text

L Natural

R Natural

ppZipperMove :: ZipperMove -> Doc a