Waargonaut.Decode.Types

Documentation

type ParseFn = ByteString -> Either DecodeError Json

type SuccinctCursor = JsonCursor ByteString Poppy512 (SimpleBalancedParens (Vector Word64))

type CursorHistory = CursorHistory' Count

newtype Decoder f a

Decoder type that is used directly to convert Json structures to other data types.

Constructors

Decoder

Fields

runDecoder :: ParseFn -> JCurs -> DecodeResultT Count DecodeError f a

newtype DecodeResult f a

Provide some of the type parameters that the underlying DecodeResultT requires. This contains the state and error management as we walk around our zipper and decode our JSON input.

Addtionally we keep our parsing function in a ReaderT such that it's accessible for all of the decoding steps.

Constructors

DecodeResult

Fields

unDecodeResult :: ReaderT ParseFn (DecodeResultT Count DecodeError f) a

newtype JCurs

Wrapper type for the SuccinctCursor

Constructors

JCurs

Fields

unJCurs :: SuccinctCursor