Waargonaut.Types.CommaSep

Description

Both arrays and objects in JSON allow for an optional trailing comma on the final element. This module houses the shared types and functions that let us handle this.

Types

data CommaSeparated ws a

This type is our possibly empty comma-separated list of values. It carries information about any leading whitespace before the first element, as well as a the rest of the elements in an Elems type.

Constructors

CommaSeparated ws (Maybe (Elems ws a))

data Elems ws a

This type represents a non-empty list of elements, enforcing that the any element but the last must be followed by a trailing comma and supporting option of a final trailing comma.

Constructors

Elems

Fields

_elemsElems :: Vector (Elem Identity ws a)

_elemsLast :: Elem Maybe ws a

class HasElems c ws a | c -> ws a where

Typeclass for things that contain an Elems structure.

Minimal complete definition

elems

Methods

elems :: Lens' c (Elems ws a)

elemsElems :: Lens' c (Vector (Elem Identity ws a))

elemsLast :: Lens' c (Elem Maybe ws a)

data Elem f ws a

Data type to represent a single element in a CommaSeparated list. Carries information about it's own trailing whitespace. Denoted by the f.

Constructors

Elem

Fields

_elemVal :: a

_elemTrailing :: f (Comma, ws)

class HasElem c f ws a | c -> f ws a where

Typeclass for things that contain a single Elem structure.

Minimal complete definition

elem

Methods

elem :: Lens' c (Elem f ws a)

elemTrailing :: Lens' c (f (Comma, ws))

elemVal :: Lens' c a

data Comma

Unary type to represent a comma.

Constructors

Comma

Parse / Build

parseComma :: CharParsing f => f Comma

Parse a single comma (,)

commaBuilder :: Builder

Builder for UTF8 Comma

parseCommaSeparated :: (Monad f, CharParsing f) => f open -> f close -> f ws -> f a -> f (CommaSeparated ws a)

Parse a CommaSeparated data structure.

>>> testparse (parseCommaSeparated (char '[') (char ']') parseWhitespace charWS) "[]"
Right (CommaSeparated (WS []) Nothing)

>>> testparse (parseCommaSeparated (char '[') (char ']') parseWhitespace charWS) "[ ]"
Right (CommaSeparated (WS [Space]) Nothing)

>>> isLeft $ testparse (parseCommaSeparated (char '[') (char ']') parseWhitespace charWS) "[ , ]"
True

>>> isLeft $ testparse (parseCommaSeparated (char '[') (char ']') parseWhitespace charWS) "[ , a]"
True

>>> isLeft $ testparse (parseCommaSeparated (char '[') (char ']') parseWhitespace charWS) "[d a]"
True

>>> testparse (parseCommaSeparated (char '[') (char ']') parseWhitespace charWS) "[d , ]"
Right (CommaSeparated (WS []) (Just (Elems {_elemsElems = [], _elemsLast = Elem {_elemVal = ('d',WS [Space]), _elemTrailing = Just (Comma,WS [Space])}})))

>>> testparse (parseCommaSeparated (char '[') (char ']') parseWhitespace charWS) "[\na\n , b]"
Right (CommaSeparated (WS [NewLine]) (Just (Elems {_elemsElems = [Elem {_elemVal = ('a',WS [NewLine,Space]), _elemTrailing = Identity (Comma,WS [Space])}], _elemsLast = Elem {_elemVal = ('b',WS []), _elemTrailing = Nothing}})))

>>> testparse (parseCommaSeparated (char '[') (char ']') parseWhitespace charWS) "[\na\n , b, \n]"
Right (CommaSeparated (WS [NewLine]) (Just (Elems {_elemsElems = [Elem {_elemVal = ('a',WS [NewLine,Space]), _elemTrailing = Identity (Comma,WS [Space])}], _elemsLast = Elem {_elemVal = ('b',WS []), _elemTrailing = Just (Comma,WS [Space,NewLine])}})))

commaSeparatedBuilder :: forall ws a. Char -> Char -> (ws -> Builder) -> (a -> Builder) -> CommaSeparated ws a -> Builder

Using the given builders for the whitespace and elements (a), create a builder for a CommaSeparated.

Conversion

_CommaSeparated :: Iso (CommaSeparated ws a) (CommaSeparated ws' b) (ws, Maybe (Elems ws a)) (ws', Maybe (Elems ws' b))

Isomorphism between the internal pieces of a CommaSeparated element.

toList :: CommaSeparated ws a -> [a]

Convert a CommaSeparated of a to [a], discarding whitespace.

fromList :: (Monoid ws, Semigroup ws) => [a] -> CommaSeparated ws a

Convert a list of a to a CommaSeparated list, with no whitespace.

Cons / Uncons

consCommaSep :: Monoid ws => ((Comma, ws), a) -> CommaSeparated ws a -> CommaSeparated ws a

Cons elements onto a CommaSeparated with provided whitespace information. If you don't need explicit whitespace then the Cons instance is more straightforward.

unconsCommaSep :: Monoid ws => CommaSeparated ws a -> Maybe ((Maybe (Comma, ws), a), CommaSeparated ws a)

Attempt to "uncons" elements from the front of a CommaSeparated without discarding the elements' whitespace information. If you don't need explicit whitespace then the Cons instance is more straightforward.