Waargonaut.Types.JArray

Description

JSON Array representation and functions.

Types

newtype JArray ws a

Conveniently, a JSON array is a CommaSeparated list with an optional trailing comma, some instances and other functions need to work differently so we wrap it up in a newtype.

Constructors

JArray (CommaSeparated ws a)

Parser / Builder

parseJArray :: (Monad f, CharParsing f) => f ws -> f a -> f (JArray ws a)

Parse a single JSON array

>>> testparse (parseJArray parseWhitespace parseWaargonaut) "[null ]"
Right (JArray (CommaSeparated (WS []) (Just (Elems {_elemsElems = [], _elemsLast = Elem {_elemVal = Json (JNull (WS [Space])), _elemTrailing = Nothing}}))))

>>> testparse (parseJArray parseWhitespace parseWaargonaut) "[null,]"
Right (JArray (CommaSeparated (WS []) (Just (Elems {_elemsElems = [], _elemsLast = Elem {_elemVal = Json (JNull (WS [])), _elemTrailing = Just (Comma,WS [])}}))))

jArrayBuilder :: (ws -> Builder) -> ((ws -> Builder) -> a -> Builder) -> JArray ws a -> Builder

Using the given builders, build a JArray.