Waargonaut.Types.JChar

Description

Types and functions for handling characters in JSON.

Types

data HexDigit4 d

JSON Characters may be single escaped UTF16 "uab34".

Constructors

HexDigit4 d d d d

class HasHexDigit4 c d | c -> d where

Typeclass for things that contain a HexDigit4.

Minimal complete definition

hexDigit4

Methods

hexDigit4 :: Lens' c (HexDigit4 d)

data JChar digit

A JChar may be unescaped or escaped.

Constructors

EscapedJChar (JCharEscaped digit)

UnescapedJChar JCharUnescaped

class AsJChar r digit | r -> digit where

Typeclass for things that be used as a JChar.

Minimal complete definition

_JChar

Methods

_JChar :: Prism' r (JChar digit)

_EscapedJChar :: Prism' r (JCharEscaped digit)

_UnescapedJChar :: Prism' r JCharUnescaped

class HasJChar c digit | c -> digit where

Typeclass for things that have a JChar.

Minimal complete definition

jChar

Methods

jChar :: Lens' c (JChar digit)

data JCharEscaped digit

Things that may be escaped in a JSON string.

Constructors

QuotationMark

ReverseSolidus

Solidus

Backspace

WhiteSpace Whitespace

Hex (HexDigit4 digit)

class AsJCharEscaped r digit | r -> digit where

Typeclass for things that may be used as an escaped JChar.

Minimal complete definition

_JCharEscaped

Methods

_JCharEscaped :: Prism' r (JCharEscaped digit)

_QuotationMark :: Prism' r ()

_ReverseSolidus :: Prism' r ()

_Solidus :: Prism' r ()

_Backspace :: Prism' r ()

_WhiteSpace :: Prism' r Whitespace

_Hex :: Prism' r (HexDigit4 digit)

newtype JCharUnescaped

Type to specify that this character is unescaped and may be represented using a normal Haskell Char.

Constructors

JCharUnescaped Char

class AsJCharUnescaped a where

Typeclass for things that may used as an unescaped JChar.

Minimal complete definition

_JCharUnescaped

Methods

_JCharUnescaped :: Prism' a JCharUnescaped

Parser / Builder

parseJChar :: (CharParsing f, HeXaDeCiMaL digit) => f (JChar digit)

Parse a JSON character.

>>> testparse parseJChar "\\u1234" :: Either DecodeError (JChar HeXDigit)
Right (EscapedJChar (Hex (HexDigit4 HeXDigit1 HeXDigit2 HeXDigit3 HeXDigit4)))

>>> testparse parseJChar "\\\\" :: Either DecodeError (JChar HeXDigit)
Right (EscapedJChar ReverseSolidus)

>>> testparse parseJChar "\\r"
Right (EscapedJChar (WhiteSpace CarriageReturn))

>>> testparsetheneof parseJChar "a"
Right (UnescapedJChar (JCharUnescaped 'a'))

>>> testparsethennoteof parseJChar "ax"
Right (UnescapedJChar (JCharUnescaped 'a'))

parseJCharEscaped :: (CharParsing f, HeXaDeCiMaL digit) => f (JCharEscaped digit)

Parse an escapted JSON character.

>>> testparse parseJCharEscaped "\\\""
Right QuotationMark

>>> testparse parseJCharEscaped "\\\\"
Right ReverseSolidus

>>> testparse parseJCharEscaped "\\/"
Right Solidus

>>> testparse parseJCharEscaped "\\b"
Right Backspace

>>> testparse parseJCharEscaped "\\f"
Right (WhiteSpace LineFeed)

>>> testparse parseJCharEscaped "\\n"
Right (WhiteSpace NewLine)

>>> testparse parseJCharEscaped "\\r"
Right (WhiteSpace CarriageReturn)

>>> testparse parseJCharEscaped "\\t"
Right (WhiteSpace HorizontalTab)

>>> testparse parseJCharEscaped "\\u1234" :: Either DecodeError (JCharEscaped HeXDigit)
Right (Hex (HexDigit4 HeXDigit1 HeXDigit2 HeXDigit3 HeXDigit4))

>>> testparsetheneof parseJCharEscaped "\\t"
Right (WhiteSpace HorizontalTab)

>>> testparsethennoteof parseJCharEscaped "\\tx"
Right (WhiteSpace HorizontalTab)

parseJCharUnescaped :: CharParsing f => f JCharUnescaped

Parse an unescaped JSON character.

>>> testparse parseJCharUnescaped "a"
Right (JCharUnescaped 'a')

>>> testparse parseJCharUnescaped "\8728"
Right (JCharUnescaped '\8728')

>>> testparsetheneof parseJCharUnescaped "a"
Right (JCharUnescaped 'a')

>>> testparsethennoteof parseJCharUnescaped "ax"
Right (JCharUnescaped 'a')

jCharBuilder :: HeXaDeCiMaL digit => JChar digit -> Builder

Create a Builder for the given JChar.

jCharToChar :: HeXaDeCiMaL digit => JChar digit -> Char

Convert a JChar to a Haskell Char.

Conversion

utf8CharToJChar :: Char -> JChar HeXDigit

Convert a Char to 'JChar HexDigit' and replace any invalid values with U+FFFD as per the Text documentation.

Refer to 'Text' documentation for more info.

jCharToUtf8Char :: JChar HeXDigit -> Maybe Char

Try to convert a JChar to a Haskell Char.