Waargonaut.Types.JObject

Description

Types and functions for handling our representation of a JSON object.

Object Type

newtype JObject ws a

The representation of a JSON object.

The JSON RFC8259 indicates that names within an object "should" be unique. But the standard does not enforce this, leaving it to the various implementations to decide how to handle it.

As there are multiple possibilities for deciding which key to use when enforcing uniqueness, Waargonaut accepts duplicate keys, allowing you to decide how to handle it.

This type is the "list of tuples of key and value" structure, as such it is a wrapper around the CommaSeparated data type.

Constructors

JObject (CommaSeparated ws (JAssoc ws a))

class HasJObject c ws a | c -> ws a where

Type class to represent something that has a JObject within it.

Minimal complete definition

jObject

Methods

jObject :: Lens' c (JObject ws a)

Key/value pair type

data JAssoc ws a

This type represents the key-value pair inside of a JSON object.

It is built like this so that we can preserve any whitespace information that may surround it.

Constructors

JAssoc

Fields

_jsonAssocKey :: JString

_jsonAssocKeyTrailingWS :: ws

_jsonAssocValPreceedingWS :: ws

_jsonAssocVal :: a

class HasJAssoc c ws a | c -> ws a where

This class allows you to write connective lenses for other data structures that may contain a JAssoc.

Minimal complete definition

jAssoc

Methods

jAssoc :: Lens' c (JAssoc ws a)

jsonAssocKey :: Lens' c JString

jsonAssocKeyTrailingWS :: Lens' c ws

jsonAssocVal :: Lens' c a

jsonAssocValPreceedingWS :: Lens' c ws

Map-like object representation

data MapLikeObj ws a

This is a newtype around our JObject for when we want to use the "map-like" representation of our JSON object. This data type will enforce that the first key found is treated as the desired element, and all subsequent occurrences of that key are discarded.

toMapLikeObj :: (Semigroup ws, Monoid ws) => JObject ws a -> (MapLikeObj ws a, [JAssoc ws a])

Take a JObject and produce a MapLikeObj where the first key is considered the unique value. Subsequence occurrences of that key and it's value are collected and returned as a list.

fromMapLikeObj :: MapLikeObj ws a -> JObject ws a

Access the underlying JObject.

Parser / Builder

jObjectBuilder :: (ws -> Builder) -> ((ws -> Builder) -> a -> Builder) -> JObject ws a -> Builder

Construct a Builder for an entire JObject, duplicate keys are preserved.

parseJObject :: (Monad f, CharParsing f) => f ws -> f a -> f (JObject ws a)

>>> testparse (parseJObject parseWhitespace parseWaargonaut) "{\"foo\":null }"
Right (JObject (CommaSeparated (WS []) (Just (Elems {_elemsElems = [], _elemsLast = Elem {_elemVal = JAssoc {_jsonAssocKey = JString' [UnescapedJChar (JCharUnescaped 'f'),UnescapedJChar (JCharUnescaped 'o'),UnescapedJChar (JCharUnescaped 'o')], _jsonAssocKeyTrailingWS = WS [], _jsonAssocValPreceedingWS = WS [], _jsonAssocVal = Json (JNull (WS [Space]))}, _elemTrailing = Nothing}}))))

>>> testparse (parseJObject parseWhitespace parseWaargonaut) "{\"foo\":null, }"
Right (JObject (CommaSeparated (WS []) (Just (Elems {_elemsElems = [], _elemsLast = Elem {_elemVal = JAssoc {_jsonAssocKey = JString' [UnescapedJChar (JCharUnescaped 'f'),UnescapedJChar (JCharUnescaped 'o'),UnescapedJChar (JCharUnescaped 'o')], _jsonAssocKeyTrailingWS = WS [], _jsonAssocValPreceedingWS = WS [], _jsonAssocVal = Json (JNull (WS []))}, _elemTrailing = Just (Comma,WS [Space])}}))))