Waargonaut.Types.Json

Description

Top level types and functions for Waargonaut Json types.

Inner JSON types

data JType ws a

Individual JSON Types and their trailing whitespace.

Constructors

JNull ws

JBool Bool ws

JNum JNumber ws

JStr JString ws

JArr (JArray ws a) ws

JObj (JObject ws a) ws

class AsJType r ws a | r -> ws a where

Typeclass for things that can represent a JType

Minimal complete definition

_JType

Methods

_JType :: Prism' r (JType ws a)

_JNull :: Prism' r ws

_JBool :: Prism' r (Bool, ws)

_JNum :: Prism' r (JNumber, ws)

_JStr :: Prism' r (JString, ws)

_JArr :: Prism' r (JArray ws a, ws)

_JObj :: Prism' r (JObject ws a, ws)

Top level JSON type

newtype Json

Top level Json type, we specialise the whitespace to WS and the digit type to Digit. Also defining that our structures can recursively only contain Json types.

Constructors

Json (JType WS Json)

Parser / Builder

waargonautBuilder :: (WS -> Builder) -> Json -> Builder

Using the given whitespace builder, create a builder for a given Json value.

parseWaargonaut :: (Monad f, CharParsing f) => f Json

Parse to a Json value, keeping all of the information about the leading and trailing whitespace.

Traversals

jsonTraversal :: Traversal' Json Json

Ignoring whitespace, traverse a Json structure.

jsonWSTraversal :: Traversal Json Json WS WS

Traverse the trailing whitespace of this Json structure.

jtypeTraversal :: Traversal (JType ws a) (JType ws a') a a'

jtypeWSTraversal :: Traversal (JType ws a) (JType ws' a) ws ws'

Traverse all of the whitespace of this Json structure and every element in the tree.