Waargonaut.Types.Whitespace

Description

Parsers and builders for whitespace characters in our JSON.

Documentation

data Whitespace

Represent the different types of whitespace.

Constructors

Space

HorizontalTab

LineFeed

NewLine

CarriageReturn

newtype WS

This is a wrapper for a sequence of consecutive whitespace.

Constructors

WS (Vector Whitespace)

_WhitespaceChar :: Prism' Char Whitespace

Handy Prism' between a Char its possible Whitespace representation.

escapedWhitespaceChar :: Whitespace -> Char

Change a Whitespace into its escaped Char form.

unescapedWhitespaceChar :: Whitespace -> Char

Change a Whitespace into a single unescaped Char. Useful if you're already handling escaping with some other mechanism.

oneWhitespace :: CharParsing f => f Whitespace

Parse a single Whitespace character.

parseWhitespace :: CharParsing f => f WS

>>> testparse parseWhitespace " "
Right (WS [Space])

>>> testparse parseWhitespace "\n    "
Right (WS [NewLine,Space,Space,Space,Space])

>>> testparse parseWhitespace " \t"
Right (WS [Space,HorizontalTab])

>>> testparse parseWhitespace "\f\f"
Right (WS [LineFeed,LineFeed])

>>> testparse parseWhitespace "\r\r\r"
Right (WS [CarriageReturn,CarriageReturn,CarriageReturn])

>>> testparse parseWhitespace "\n\r\r\n"
Right (WS [NewLine,CarriageReturn,CarriageReturn,NewLine])

>>> testparse parseWhitespace ""
Right (WS [])

>>> testparse parseWhitespace "\n   ]"
Right (WS [NewLine,Space,Space,Space])

parseSomeWhitespace :: CharParsing f => f (NonEmpty Whitespace)

Parse a NonEmpty sequence of consecutive whitespace.

wsBuilder :: WS -> Builder

Reconstitute the given whitespace into its original form.

wsRemover :: WS -> Builder

Remove any whitespace. Minification for free, yay!