Waargonaut.Generic

TypeClasses

class JsonEncode t a where

Methods

mkEncoder :: Applicative f => Tagged t (Encoder f a)

mkEncoder :: (Applicative f, Generic a, HasDatatypeInfo a, All2 (JsonEncode t) (Code a)) => Tagged t (Encoder f a)

class JsonDecode t a where

Methods

mkDecoder :: Monad f => Tagged t (Decoder f a)

mkDecoder :: (Monad f, Generic a, HasDatatypeInfo a, All2 (JsonDecode t) (Code a)) => Tagged t (Decoder f a)

Tag

data GWaarg

This is a provided tag that can be used for tagging the JsonEncode and JsonDecode instances.

Options

data NewtypeName

The two options we currently have for using the Generic mechanism to handle 'newtype' values:

Constructors

Unwrap
Discard the newtype wrapper and encode the inner value.

newtype Foo = Foo Text

let x = Foo Fred

Will be encoded as: Fred

ConstructorNameAsKey
Encode the newtype value as an object using the constructor as the "key".

newtype Foo = Foo Text

let x = Foo Fred

Will be encoded as: "{"foo":"Fred"}"

data Options

The configuration options for creating Generic encoder or decoder values.

Constructors

Options

Fields

_optionsFieldName :: String -> String

When encoding/decoding a record type, this function will be used on the field names to determine how they will be encoded. Or what keys to look up on the JSON object when it is being decoded.

_optionsNewtypeWithConsName :: NewtypeName

How to handle 'newtype' values. See NewtypeName for more info.

defaultOpts :: Options

Default options for Generic functionality:

Field names are left untouched (id)

Newtype values are encoded as raw values (Unwrap)

Creation

gEncoder :: forall t a f. (Generic a, Applicative f, HasDatatypeInfo a, All2 (JsonEncode t) (Code a)) => Options -> Tagged t (Encoder f a)

gDecoder :: forall f a t. (Generic a, HasDatatypeInfo a, All2 (JsonDecode t) (Code a), Monad f) => Options -> Tagged t (Decoder f a)

Reexports

module Data.Tagged

class All (SListI :: [*] -> Constraint) (Code a) => Generic a where

The class of representable datatypes.

The SOP approach to generic programming is based on viewing datatypes as a representation (Rep) built from the sum of products of its components. The components of are datatype are specified using the Code type family.

The isomorphism between the original Haskell datatype and its representation is witnessed by the methods of this class, from and to. So for instances of this class, the following laws should (in general) hold:

to . from === id :: a -> a
from . to === id :: Rep a -> Rep a

You typically don't define instances of this class by hand, but rather derive the class instance automatically.

Option 1: Derive via the built-in GHC-generics. For this, you need to use the DeriveGeneric extension to first derive an instance of the Generic class from module GHC.Generics. With this, you can then give an empty instance for Generic, and the default definitions will just work. The pattern looks as follows:

import qualified GHC.Generics as GHC
import Generics.SOP

...

data T = ... deriving (GHC.Generic, ...)

instance Generic T -- empty
instance HasDatatypeInfo T -- empty, if you want/need metadata

Option 2: Derive via Template Haskell. For this, you need to enable the TemplateHaskell extension. You can then use deriveGeneric from module Generics.SOP.TH to have the instance generated for you. The pattern looks as follows:

import Generics.SOP
import Generics.SOP.TH

...

data T = ...

deriveGeneric ''T -- derives HasDatatypeInfo as well

Tradeoffs: Whether to use Option 1 or 2 is mainly a matter of personal taste. The version based on Template Haskell probably has less run-time overhead.

Non-standard instances: It is possible to give Generic instances manually that deviate from the standard scheme, as long as at least

to . from === id :: a -> a

still holds.

Associated Types

type Code a :: [[*]]

The code of a datatype.

This is a list of lists of its components. The outer list contains one element per constructor. The inner list contains one element per constructor argument (field).

Example: The datatype

data Tree = Leaf Int | Node Tree Tree

is supposed to have the following code:

type instance Code (Tree a) =
  '[ '[ Int ]
   , '[ Tree, Tree ]
   ]

Methods

from :: a -> Rep a

Converts from a value to its structural representation.

to :: Rep a -> a

Converts from a structural representation back to the original value.

class HasDatatypeInfo a where

A class of datatypes that have associated metadata.

It is possible to use the sum-of-products approach to generic programming without metadata. If you need metadata in a function, an additional constraint on this class is in order.

You typically don't define instances of this class by hand, but rather derive the class instance automatically. See the documentation of Generic for the options.

Associated Types

type DatatypeInfoOf a :: DatatypeInfo

Type-level datatype info

Methods

datatypeInfo :: proxy a -> DatatypeInfo (Code a)

Term-level datatype info; by default, the term-level datatype info is produced from the type-level info.